0520 added permissions and increased payload size to 1mb

+++ api/guild/guild.js
@@ -19,6 +19,9 @@ export async function getCommands(req, res) {
         if(session && !(await DB.sessionInGuild(session, guildId))) {
             throw new MyError(401, 'Session Id is not present in guild')
         }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
+        }
         return DB.getCommands(guildId, type, projection)
     })
 }
@@ -31,6 +34,9 @@ export async function getCommand(req, res) {
     processRequest(res, async () => {
         if(session && !(await DB.sessionInGuild(session, guildId))) {
             throw new MyError(401, 'Session Id is not present in guild')
+        }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
         }
         return DB.getCommand(commandId, projection)
     })
@@ -47,6 +53,9 @@ export async function addCommand(req, res) {
         if(session && !(await DB.sessionIsGuildOfficer(session, guildId))) {
             throw new MyError(401, 'Session Id is not a guild officer')
         }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
+        }
         return DB.addCommand(commandId, type, title, description, guildId)
     })
 }
@@ -58,7 +67,10 @@ export async function deleteCommand(req, res) {
     processRequest(res, async () => {
         if(session && !(await DB.sessionIsGuildOfficer(session, guildId))) {
             throw new MyError(401, 'Session Id is not a guild officer')
-        } 
+        }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
+        }
         return DB.deleteCommand(commandId)
     })
 }
@@ -70,6 +82,9 @@ export async function getOperations(req, res) {
     processRequest(res, async () => {
         if(session && !(await DB.sessionInGuild(session, guildId))) {
             throw new MyError(401, 'Session Id is not present in guild')
+        }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
         }
         return DB.getOperations(guildId, projection)
     })
@@ -84,6 +99,9 @@ export async function getOperation(req, res) {
         if(session && !(await DB.sessionInGuild(session, guildId))) {
             throw new MyError(401, 'Session Id is not present in guild')
         }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
+        }
         return DB.getOperation(operationId, projection)
     })
 }
@@ -94,8 +112,11 @@ export async function addOperation(req, res) {
     let operationId = req.body.operationId
     let operation = req.body.operation
     processRequest(res, async () => {
-        if(session && !(await DB.sessionInGuild(session, guildId))) {
+        if(session && !(await DB.sessionIsGuildOfficer(session, guildId))) {
             throw new MyError(401, 'Session Id is not a guild officer')
+        }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
         }
         return DB.addOperation(operation, operationId, guildId)
     })
@@ -106,9 +127,23 @@ export async function deleteOperation(req, res) {
     let session = req.body.session
     let operationId = req.body.operationId
     processRequest(res, async () => {
+        if(session && !(await DB.sessionIsGuildOfficer(session, guildId))) {
+            throw new MyError(401, 'Session Id is not a guild officer')
+        }
+        if(!DB.isGuildBuild(guildId)) {
+            throw new MyError(401, 'Guild is not registered with the guild build.')
+        }
+        return DB.deleteOperation(operationId)
+    })
+}
+
+export async function isGuildBuild(req, res) {
+    let guildId = req.body.guildId
+    let session = req.body.session
+    processRequest(res, async () => {
         if(session && !(await DB.sessionInGuild(session, guildId))) {
             throw new MyError(401, 'Session Id is not a guild officer')
         } 
-        return DB.deleteOperation(operationId)
+        return DB.isGuildBuild(guildId)
     })
 }
